+++ examples/editor/src/Plugins.jsx
@@ -33,6 +33,7 @@ import 'wix-rich-content-plugin-giphy/dist/styles.min.css';
 import 'wix-rich-content-plugin-youtube/dist/styles.min.css';
 import 'wix-rich-content-plugin-map/dist/styles.min.css';
 import 'wix-rich-content-plugin-file-upload/dist/styles.min.css';
+import 'wix-rich-content-plugin-text-color/dist/styles.min.css';
 
 const plugins = [
   createImagePlugin,