+++ packages/plugin-text-color/src/toolbar/TextColorPanel.jsx
@@ -2,15 +2,30 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { Modifier, EditorState } from '@wix/draft-js';
 import { ColorPicker, getSelectionStyles } from 'wix-rich-content-common';
-import { isHexColor } from '../utils';
-import { DEFAULT_PALETTE, DEFAULT_COLOR, DEFAULT_SELECTION_COLOR } from './constants';
+import { DEFAULT_COLOR, DEFAULT_STYLE_SELECTION_PREDICATE } from '../constants';
+
+import {
+  extractColor,
+  extractPalette,
+  extractSchemeAttributes,
+  validateColorScheme,
+} from '../color-scheme-utils';
 
 export default class TextColorPanel extends Component {
   constructor(props) {
     super(props);
-    const currentColors = getSelectionStyles(style => isHexColor(style), props.editorState);
+    const styleSelectionPredicate =
+      props.settings.styleSelectionPredicate || DEFAULT_STYLE_SELECTION_PREDICATE;
+    if (props.settings.colorScheme && !validateColorScheme(props.settings.colorScheme)) {
+      console.error('Error: colorScheme is not valid'); // eslint-disable-line no-console
+    }
+    const currentColors = getSelectionStyles(styleSelectionPredicate, props.editorState);
     this.state = {
-      currentColor: currentColors.length > 0 ? currentColors[0] : DEFAULT_COLOR,
+      currentColor:
+        currentColors.length > 0
+          ? extractColor(props.settings.colorScheme, currentColors[0])
+          : DEFAULT_COLOR,
+      currentSchemeColor: currentColors[0],
       userColors: props.settings.getUserColors() || [],
     };
     this.setColor = this.setColor.bind(this);
@@ -22,29 +37,31 @@ export default class TextColorPanel extends Component {
   }
 
   setColor(color) {
+    let { editorState } = this.props;
     if (color !== this.state.currentColor) {
-      this.applyInlineColorStyle(color);
+      editorState = this.getInlineColorState(color);
       this.setState({
-        currentColor: color,
+        currentColor: extractColor(this.props.settings.colorScheme, color),
+        currentSchemeColor: color,
       });
     }
-    this.props.helpers.closeModal();
+    this.props.closeModal(editorState);
   }
 
-  applyInlineColorStyle(color) {
-    const { editorState, setEditorState } = this.props;
-    const currentColors = getSelectionStyles(style => isHexColor(style), editorState);
+  getInlineColorState(color) {
+    const { editorState, settings } = this.props;
+    const styleSelectionPredicate =
+      settings.styleSelectionPredicate || DEFAULT_STYLE_SELECTION_PREDICATE;
+    const selection = editorState.getSelection();
+    const currentColors = getSelectionStyles(styleSelectionPredicate, editorState);
     const newEditorState = currentColors.reduce((nextEditorState, prevColor) => {
-      const selection = nextEditorState.getSelection();
       const contentState = nextEditorState.getCurrentContent();
       const nextContentState = Modifier.removeInlineStyle(contentState, selection, prevColor);
       return EditorState.push(nextEditorState, nextContentState, 'change-inline-style');
     }, editorState);
-
-    const selection = newEditorState.getSelection();
     const contentState = newEditorState.getCurrentContent();
     const newContentState = Modifier.applyInlineStyle(contentState, selection, color);
-    setEditorState(EditorState.push(newEditorState, newContentState, 'change-inline-style'));
+    return EditorState.push(newEditorState, newContentState, 'change-inline-style');
   }
 
   onColorAdded(color) {
@@ -55,19 +72,26 @@ export default class TextColorPanel extends Component {
   }
 
   render() {
-    const { theme, settings, t, setKeepToolbarOpen } = this.props;
-    const palette = settings.palette || DEFAULT_PALETTE;
+    const { theme, settings, t, setKeepToolbarOpen, isMobile } = this.props;
+    const { colorScheme } = settings;
+    const palette = extractPalette(colorScheme);
+    const schemeAttributes = extractSchemeAttributes(colorScheme);
+    const { onCustomPickerToggle, onCustomColorPicked } = settings;
     return (
       <ColorPicker
+        schemeAttributes={schemeAttributes}
+        schemeColor={this.state.currentSchemeColor}
         color={this.state.currentColor}
-        selectionColor={settings.selectionColor || DEFAULT_SELECTION_COLOR}
         palette={palette.slice(0, 6)}
         userColors={this.state.userColors.slice(0, 17)}
         onColorAdded={this.onColorAdded}
         onChange={this.setColor}
+        onCustomPickerToggle={onCustomPickerToggle}
+        onCustomColorPicked={onCustomColorPicked}
         theme={theme}
         t={t}
         setKeepToolbarOpen={setKeepToolbarOpen}
+        isMobile={isMobile}
       />
     );
   }
@@ -81,10 +105,13 @@ TextColorPanel.propTypes = {
   uiSettings: PropTypes.object,
   settings: PropTypes.shape({
     onColorAdded: PropTypes.func.isRequired,
-    palette: PropTypes.arrayOf(PropTypes.string),
+    colorScheme: PropTypes.object,
     getUserColors: PropTypes.func,
-    selectionColor: PropTypes.string,
+    onCustomPickerToggle: PropTypes.func,
+    onCustomColorPicked: PropTypes.func,
+    styleSelectionPredicate: PropTypes.func,
   }).isRequired,
   setKeepToolbarOpen: PropTypes.func,
-  helpers: PropTypes.object.isRequired,
+  closeModal: PropTypes.func.isRequired,
+  isMobile: PropTypes.bool.isRequired,
 };